+++ webview-ui/src/components/chat/BatchDiffApproval.tsx
@@ -1,5 +1,5 @@
 import React, { memo, useState } from "react"
-import CodeAccordian from "../common/CodeAccordian"
+import CodeAccordion from "../common/CodeAccordion"
 
 interface FileDiff {
 	path: string
@@ -41,7 +41,7 @@ export const BatchDiffApproval = memo(({ files = [], ts }: BatchDiffApprovalProp
 
 					return (
 						<div key={`${file.path}-${index}-${ts}`}>
-							<CodeAccordian
+							<CodeAccordion
 								path={file.path}
 								code={unified}
 								language="diff"